Abstract the ComponentRegistry
