+++ lib/streamlit/web/server/component_request_handler.py
@@ -21,14 +21,14 @@ from typing import Final
 import tornado.web
 
 import streamlit.web.server.routes
-from streamlit.components.v1.components import ComponentRegistry
+from streamlit.components.v1.base_component_registry import BaseComponentRegistry
 from streamlit.logger import get_logger
 
 _LOGGER: Final = get_logger(__name__)
 
 
 class ComponentRequestHandler(tornado.web.RequestHandler):
-    def initialize(self, registry: ComponentRegistry):
+    def initialize(self, registry: BaseComponentRegistry):
         self._registry = registry
 
     def get(self, path: str) -> None:
